mqtt/consumer/requirements.txt
setuptools
paho-mqtt>=1.4.0
elasticsearch>=7.0.5